factor out feature vector tests to their own module create separate test script for ROI functionality
